release/EmoScan-v1.0.0-Developer/backend-service/pyproject.toml
[project]
name = "emoscan-backend"
version = "1.0.0"
description = "EmoScan情感分析后端API服务"
authors = [
    {name = "EmoScan Team"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pillow>=10.1.0",
    "openai>=1.3.7",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]